+++ 1_2d_game/src/main.cpp
@@ -4,6 +4,8 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
 #include <iostream>
+#include <variant>
+#include <vector>
 #include "collision.hpp"
 
 /// @brief Interface for objects that can be drawn
@@ -14,6 +16,8 @@ struct Drawable {
     virtual ~Drawable() = default;
 };
 
+bool keyStates[256] = {false};
+
 /// @brief Interface for objects that can be updated
 struct Updatable {
     /// @brief Update the object's state. Return true if the object should be removed.
@@ -23,6 +27,93 @@ struct Updatable {
     virtual ~Updatable() = default;
 };
 
+struct EnemyBullet : Updatable, Drawable, Collidable {
+    glm::fvec2 initialDirection;
+    glm::fvec2 initialPosition;
+
+    EnemyBullet(glm::fvec2 initialDirection, glm::fvec2 initialPosition)
+        : initialDirection(initialDirection), initialPosition(initialPosition) {}
+    ~EnemyBullet() {}
+
+    bool update(int t) override { return false; }
+    void draw(glm::fvec2 cameraOffset) override {}
+    CollisionShape getShape() const override { return CollisionCircle(glm::vec2(0.0, 0.0), 1.0); }
+};
+
+struct PlayerBullet : Updatable, Drawable, Collidable {
+    glm::fvec2 initialPosition;
+
+    PlayerBullet(glm::fvec2 initialPosition) : initialPosition(initialPosition) {}
+    ~PlayerBullet() {}
+
+    bool update(int t) override { return false; }
+    void draw(glm::fvec2 cameraOffset) override {}
+    CollisionShape getShape() const override { return CollisionCircle(glm::vec2(0.0, 0.0), 1.0); }
+};
+
+struct Player : Updatable, Drawable, Collidable {
+    glm::fvec2 currentPosition;
+
+    Player(glm::fvec2 initialPosition) : currentPosition(initialPosition) {}
+    ~Player() {}
+
+    bool update(int t) override { return false; }
+    void draw(glm::fvec2 cameraOffset) override {}
+    CollisionShape getShape() const override { return CollisionCircle(glm::vec2(0.0, 0.0), 1.0); }
+};
+
+struct Boss : Updatable, Drawable, Collidable {
+    glm::fvec2 currentPosition;
+    
+    Boss(glm::fvec2 initialPosition) : currentPosition(initialPosition) {}
+    ~Boss() {}
+
+    bool update(int t) override { return false; }
+    void draw(glm::fvec2 cameraOffset) override {}
+    CollisionShape getShape() const override { return CollisionCircle(glm::vec2(0.0, 0.0), 1.0); }
+};
+
+struct Hearts : Drawable {
+    glm::fvec2 drawPosition;
+
+    Hearts(glm::fvec2 drawPosition) : drawPosition(drawPosition) {}
+    ~Hearts() {}
+
+    void draw(glm::fvec2 cameraOffset) override {}
+};
+
+struct HealthBar : Drawable {
+    glm::fvec2 drawPosition;
+
+    HealthBar(glm::fvec2 drawPosition) : drawPosition(drawPosition) {}
+    ~HealthBar() {}
+
+    void draw(glm::fvec2 cameraOffset) override {}
+};
+
+using Object = std::variant<Drawable *, Collidable *, Updatable *>;
+
+struct GameState {
+    GameState(int h, int bh)
+        : health(h), bossHealth(bh), cameraOffset(0.0f, 0.0f), playerSprite(glm::fvec2(0.0f, 0.0f)),
+          bossSprite(glm::fvec2(0.0f, 0.0f)), healthBarSprite(glm::fvec2(0.0f, 0.0f)), heartsSprite(glm::fvec2(0.0f, 0.0f)) {}
+
+    int health;
+    int bossHealth;
+    glm::fvec2 cameraOffset;
+
+    Player playerSprite;
+    Boss bossSprite;
+    HealthBar healthBarSprite;
+    Hearts heartsSprite;
+
+    std::vector<PlayerBullet> playerBulletSprites;
+std::vector<EnemyBullet> enemyBulletSprites;
+};
+
+void keyboardDown(unsigned char key, int x, int y) { keyStates[key] = true; }
+void keyboardUp(unsigned char key, int x, int y) { keyStates[key] = false; }
+
 void display() {
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
@@ -30,6 +121,19 @@ void display() {
 
     glutSwapBuffers();
     glutPostRedisplay();
+}
+
+void update() {
+    if (keyStates[27]) {
+        std::cout << "ESC pressed -> exit\n";
+        std::exit(0); // End Main Loop
+    }
+    if (keyStates['a']) {
+        std::cout << "a clicked" << std::endl;
+    }
+    if (keyStates['w']) {
+        std::cout << "w clicked" << std::endl;
+    }
 }
 
 void timer(int value) {
@@ -41,7 +145,7 @@ int main(int argc, char **argv) {
     glutInit(&argc, argv);
     glutInitDisplayMode(GLUT_DOUBLE | GLUT_RGBA | GLUT_DEPTH);
     glutInitWindowSize(600, 600);
-    glutCreateWindow("FreeGLUT + GLEW + GLM Example");
+    glutCreateWindow("FreeGLUT + GLEW + GLM Example 2");
 
     // Test GLM properly linked
     glm::vec3 glmTest(1.0f, 0.0f, 0.0f);
@@ -52,8 +156,12 @@ int main(int argc, char **argv) {
         return -1;
     }
 
+    GameState gameState(100, 500);
+
     glEnable(GL_DEPTH_TEST);
 
+    glutKeyboardFunc(keyboardDown);
+    glutKeyboardUpFunc(keyboardUp);
     glutDisplayFunc(display);
     glutTimerFunc(0, timer, 0);
